+++ continuous_integration/travis/install.sh
@@ -69,18 +69,32 @@ create_new_conda_env() {
 
 create_new_conda_env
 
+# Install py.test timeout to fasten failure in deadlocking tests
+PIP_INSTALL_PACKAGES="pytest-timeout"
+
 if [ -n "$NUMPY_VERSION" ]; then
     # We want to ensure no memory copies are performed only when numpy is
     # installed. This also ensures that we don't keep a strong dependency on
-    # memory_profiler.
-    pip install memory_profiler
+    # memory_profiler. We also want to ensure that joblib can be used with and
+    # without lz4 compressor package installed.
+    PIP_INSTALL_PACKAGES="$PIP_INSTALL_PACKAGES memory_profiler"
+    if [ "$NO_LZ4" != "true" ]; then
+        PIP_INSTALL_PACKAGES="$PIP_INSTALL_PACKAGES lz4"
+    fi
 fi
 
-# Install py.test timeout to fasten failure in deadlocking tests
-pip install pytest-timeout
-
 if [[ "$COVERAGE" == "true" ]]; then
-    pip install pytest-cov codecov
+    PIP_INSTALL_PACKAGES="$PIP_INSTALL_PACKAGES pytest-cov codecov"
+fi
+
+pip install $PIP_INSTALL_PACKAGES
+
+
+if [[ "$CYTHON" == "true" ]]; then
+    pip install cython
+    cd joblib/test/_openmp_test_helper
+    python setup.py build_ext -i
+    cd ../../..
 fi
 
 python setup.py install